// lib/kvarteret-personal-api/types.gen.ts
export type EventDetail = {
    /**
     * Created At
     */
    created_at: string
    /**
     * Description
     */
    description?: string | null
    /**
     * Ends At
     */
    ends_at: string
    event_type?: EventType | null
    /**
     * Event Type Id
     */
    event_type_id: string
    /**
     * Facebook Url
     */
    facebook_url?: string | null
    /**
     * Id
     */
    id: string
    /**
     * Image Caption
     */
    image_caption?: string | null
    /**
     * Image Url
     */
    image_url?: string | null
    /**
     * Is Featured
     */
    is_featured: boolean
    /**
     * Is Internal
     */
    is_internal: boolean
    /**
     * Language
     */
    language: "no" | "en"
    /**
     * Organizer Groups
     */
    organizer_groups: Array<EventOrganizerGroup>
    /**
     * Price
     */
    price?: string | null
    /**
     * Recurring Interval Days
     */
    recurring_interval_days?: number | null
    room?: EventRoom | null
    /**
     * Room Id
     */
    room_id?: string | null
    /**
     * Room Text
     */
    room_text?: string | null
    /**
     * Slug
     */
    slug: string
    /**
     * Starts At
     */
    starts_at: string
    /**
     * Status
     */
    status: "published" | "draft" | "archived"
    /**
     * Ticket Url
     */
    ticket_url?: string | null
    /**
     * Title
     */
    title: string
    translations: EventTranslations
    /**
     * Updated At
     */
    updated_at: string
}

/**
 * EventList
 */
export type EventList = {
    /**
     * Events
     */
    events: Array<EventDetail>
}

/**
 * EventOrganizerGroup
 */
export type EventOrganizerGroup = {
    /**
     * Default Event Type Id
     */
    default_event_type_id?: string | null
    /**
     * Id
     */
    id: string
    /**
     * Is Active
     */
    is_active: boolean
    /**
     * Name
     */
    name: string
    /**
     * Slug
     */
    slug: string
    /**
     * Sort Order
     */
    sort_order: number
}

/**
 * EventRoom
 */
export type EventRoom = {
    /**
     * Id
     */
    id: string
    /**
     * Is Active
     */
    is_active: boolean
    /**
     * Name
     */
    name: string
    /**
     * Slug
     */
    slug: string
    /**
     * Sort Order
     */
    sort_order: number
}

/**
 * EventTaxonomy
 */
export type EventTaxonomy = {
    /**
     * Event Type Groups
     */
    event_type_groups: Array<EventTypeGroup>
    /**
     * Organizer Groups
     */
    organizer_groups: Array<EventOrganizerGroup>
    /**
     * Rooms
     */
    rooms: Array<EventRoom>
}

/**
 * EventTranslation
 */
export type EventTranslation = {
    /**
     * Available
     */
    available?: boolean
    /**
     * Description
     */
    description?: string | null
    /**
     * Image Caption
     */
    image_caption?: string | null
    /**
     * Title
     */
    title: string
}

/**
 * EventTranslations
 */
export type EventTranslations = {
    en?: EventTranslation | null
    no?: EventTranslation | null
}

/**
 * EventType
 */
export type EventType = {
    /**
     * Description
     */
    description?: string | null
    /**
     * Id
     */
    id: string
    /**
     * Is Active
     */
    is_active: boolean
    /**
     * Name
     */
    name: string
    /**
     * Slug
     */
    slug: string
    /**
     * Sort Order
     */
    sort_order: number
    /**
     * Taxonomy Group
     */
    taxonomy_group: string
}

/**
 * EventTypeGroup
 */
export type EventTypeGroup = {
    /**
     * Event Types
     */
    event_types: Array<EventType>
    /**
     * Name
     */
    name: string
}